added dockerOptions to the docu generation
